Flyttet slette og lagre ut i egen fil for bedre oversikt
